Fix crash populating design with special nets
When a Reader is reading a design and both a verilog and a def file are
used for the description of the design, the special nets are only
defined in the def and not in the verilog.

This fix allows the definition of normal nets from the verilog file, and
adds the definitions of the special nets from the def file.