Fix all reordering warnings reported by gcc with -Wreorder
During compilation of rsyn with warnings enabled, there are thousands of
warnings. This patch fixes all the warnings produced when -Wreorder is
passed to gcc (enabled by default with -Wall).

The fix moves the initialization of member variables from the constructors
to default member initializers (C++ 11), which make it easier to know the
default value, it's clearer to see which members might have not been
default-initialized, and in many cases allows the removal of the default
constructor, moving all the job to the compiler (less code == less bugs).